requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
django-phonenumber-field==7.3.0
django-resized==1.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.1.1
idna==3.7
kavenegar==1.1.2
phonenumbers==8.13.36
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1